python/0010/requirements.txt
# 0010 rely package
pillow
